adf_vol / adfCreateVol: int types clean-up.

+++ src/adf_vol.c
@@ -247,8 +247,8 @@ void adfUnMount ( struct AdfVolume * const vol )
  * 
  */
 struct AdfVolume * adfCreateVol ( struct AdfDevice * const dev,
-                                  const int32_t            start,
-                                  const int32_t            len,
+                                  const uint32_t           start,
+                                  const uint32_t           len,
                                   const char * const       volName,
                                   const uint8_t            volType )
 {
@@ -268,9 +268,9 @@ struct AdfVolume * adfCreateVol ( struct AdfDevice * const dev,
     }
 	
     vol->dev = dev;
-    vol->firstBlock = (dev->heads * dev->sectors)*start;
-    vol->lastBlock = (vol->firstBlock + (dev->heads * dev->sectors)*len)-1;
-    vol->rootBlock = (vol->lastBlock - vol->firstBlock+1)/2;
+    vol->firstBlock = (int32_t) ( dev->heads * dev->sectors * start );
+    vol->lastBlock = vol->firstBlock + (int32_t) ( dev->heads * dev->sectors * len ) - 1;
+    vol->rootBlock = ( vol->lastBlock - vol->firstBlock + 1 ) / 2;
 /*printf("first=%ld last=%ld root=%ld\n",vol->firstBlock,
  vol->lastBlock, vol->rootBlock);
 */
